requirements.txt
openai==1.11.1
langchain==0.1.7
langchain-community==0.0.17
gradio
python-dotenv
discord.py
responses
tqdm
cohere
lru_cache
transformers
pytorch
qdrant-client
langchain-google-genai
pytube
openai-whisper
youtube-transcript-api